Restructures and fixes tests.
