+++ package.json
@@ -87,7 +87,7 @@
       "url": "https://github.com/strues"
     },
     {
-      "name": "Tom (datoml)",
+      "name": "Tom",
       "url": "https://github.com/datoml"
     }
   ],
