+++ shared/components/DemoApp/AsyncAbout/__tests__/About.test.js
@@ -2,7 +2,8 @@
 
 import React from 'react';
 import { shallow } from 'enzyme';
-import About from './About';
+
+import About from '../About';
 
 describe('<About />', () => {
   test('renders', () => {
