+++ shared/components/DemoApp/AsyncHome/__tests__/Home.test.js
@@ -2,7 +2,8 @@
 
 import React from 'react';
 import { shallow } from 'enzyme';
-import Home from './Home';
+
+import Home from '../Home';
 
 describe('<Home />', () => {
   test('renders', () => {
