+++ shared/components/DemoApp/Error404/__tests__/Error404.test.js
@@ -1,0 +1,17 @@
+/* eslint-disable import/no-extraneous-dependencies */
+
+import React from 'react';
+import { shallow } from 'enzyme';
+
+import Error404 from '../index';
+
+describe('<Error404 />', () => {
+  test('renders', () => {
+    const routerContext = {
+      staticContext: {},
+    };
+    const wrapper = shallow(<Error404 />, { context: { router: routerContext } });
+    expect(wrapper).toMatchSnapshot();
+    expect(routerContext.staticContext.missed).toBeTruthy();
+  });
+});
