+++ shared/components/DemoApp/Header/Menu/__tests__/Menu.test.js
@@ -2,7 +2,8 @@
 
 import React from 'react';
 import { shallow } from 'enzyme';
-import Menu from './index';
+
+import Menu from '../index';
 
 describe('<Home />', () => {
   test('renders', () => {
